backend_django/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9,<2.10
dj-database-url>=2.1,<2.3
gunicorn>=22.0,<23.0
django-cors-headers>=4.3,<4.5
